refactor(store): single file modules
