--- src/state/modules/auth/getters.js
@@ -1,7 +1,0 @@
-export function loggedIn(state) {
-  return !!state.currentUser
-}
-
-export function apiToken(state) {
-  return (state.currentUser && state.currentUser.api_token) || null
-}
